# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GolfTracker API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://backend-production-b6713.up.railway.app" },
    { "url": "http://localhost:3001" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "HoleScore": {
        "type": "object",
        "properties": {
          "id":          { "type": "integer" },
          "round_id":    { "type": "integer" },
          "hole_number": { "type": "integer", "minimum": 1, "maximum": 18 },
          "score":       { "type": "integer" },
          "par":         { "type": "integer" },
          "created_at":  { "type": "string", "format": "date-time" }
        }
      },
      "Round": {
        "type": "object",
        "properties": {
          "id":          { "type": "integer" },
          "user_id":     { "type": "integer" },
          "course_id":   { "type": "integer" },
          "date_played": { "type": "string", "format": "date" },
          "total_score": { "type": "integer" },
          "notes":       { "type": "string" },
          "course_name": { "type": "string" },
          "course_par":  { "type": "integer" },
          "first_name":  { "type": "string" },
          "last_name":   { "type": "string" },
          "created_at":  { "type": "string", "format": "date-time" }
        }
      },
      "RoundDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Round" },
          {
            "type": "object",
            "properties": {
              "email":       { "type": "string" },
              "hole_scores": { "type": "array", "items": { "$ref": "#/components/schemas/HoleScore" } }
            }
          }
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error":  { "type": "string" },
          "stack":  { "type": "string" },
          "detail": { "type": "string" },
          "query":  { "type": "string" },
          "hint":   { "type": "string" },
          "code":   { "type": "string" }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      }
    }
  },
  "paths": {
...
    "/api/rounds": {
      "get": {
        "tags": ["Rounds"],
        "summary": "List rounds",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "user_id", "in": "query", "schema": { "type": "integer" }, "description": "View another user's rounds (IDOR)" }
        ],
        "responses": {
          "200": { "description": "Round list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Round" } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "tags": ["Rounds"],
        "summary": "Create a round",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["course_id", "date_played"],
                "properties": {
                  "course_id":   { "type": "integer" },
                  "date_played": { "type": "string", "format": "date" },
                  "total_score": { "type": "integer" },
                  "notes":       { "type": "string" },
                  "user_id":     { "type": "integer", "description": "Override user (mass assignment)" },
                  "hole_scores": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "hole_number": { "type": "integer" },
                        "score":       { "type": "integer" },
                        "par":         { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created round", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RoundDetail" } } } }
        }
      }
    },
    "/api/rounds/{id}": {
      "get": {
        "tags": ["Rounds"],
        "summary": "Get a round by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Round detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RoundDetail" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "tags": ["Rounds"],
        "summary": "Update a round",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["course_id", "date_played", "total_score"],
                "properties": {
                  "course_id":   { "type": "integer" },
                  "date_played": { "type": "string", "format": "date" },
                  "total_score": { "type": "integer" },
                  "notes":       { "type": "string" },
                  "hole_scores": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "hole_number": { "type": "integer" },
                        "score":       { "type": "integer" },
                        "par":         { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated round", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RoundDetail" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Rounds"],
        "summary": "Delete a round",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } } }
        }
      }
    },
...
    "/api/admin/rounds": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all rounds",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Round list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Round" } } } } }
        }
      }
    },
    "/api/admin/rounds/{id}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a round",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } } }
        }
      }
    },
...
  }
}